refactor(graph_view): typed PrBadge + badges module (#78)
Move the PR-badge and merged-badge decisions out of graph_view/mod.rs
into a new badges.rs: pr_for_branch_labels, pr_for_row, pr_badge_text,
pr_badge_color, the PR icon consts, and merged_badge / merged_style.
These are pure over PrInfo / PrContext / Theme.

pr_for_row now returns Option<PrBadge> (typed text + color) and takes the
theme, folding the former render-time pr_badge_text/pr_badge_color calls
into the badge decision; the render tail draws badge.text / badge.color
directly. pr_for_branch_labels keeps its &PrInfo return and is re-exported
as graph_view::pr_for_branch_labels for its app-side callers.

MERGE_ICON and merged_chip_style stay in mod.rs (the collapse-merge and
merged-chip rendering, plus the #81-bound PR-subject rewrite, still live
there). Badge decision tests moved to badges.rs; the full-row approved+
comment test converted to assert on PrBadge text; ordering/layout tests
(badge before pill, separator spacing) kept as full renders in mod.rs.
